+++ src/components/reviewer/reviewer.tsx
@@ -1,12 +1,16 @@
-import { ReviewerProps } from "../../types";
+import { useSelector } from "../../services/hooks";
 import styles from "./reviewer.module.css";
 
-export const Reviewer: React.FC<ReviewerProps> = ({ login, url }) => {
+export const Reviewer = () => {
+  const { reviewer } = useSelector((store) => store.reviewer);
+  if (!reviewer) {
+    return;
+  }
   return (
     <div className={styles.container}>
       <h2>Your reviewer:</h2>
-      <p>{login}</p>
-      <a href={url}>{url}</a>
+      <p>{reviewer.login}</p>
+      <a href={reviewer.url}>{reviewer.url}</a>
     </div>
   );
 };
