+++ src/services/constants/index.ts
@@ -5,4 +5,6 @@ export const GET_REVIEWER__REQUEST = 'GET_REVIEWER__REQUEST';
 export const GET_REVIEWER__SUCCESS = 'GET_REVIEWER__SUCCESS';
 export const GET_REVIEWER__FAILURE = 'GET_REVIEWER__SUCCESS';
 
-export const UPDATE__FROM_LS = 'UPDATE__FROM_LS'
+export const UPDATE_FROM__LS = 'UPDATE_FROM__LS';
+
+export const UPDATE_FROM__STATE = 'UPDATE_FROM__STATE';
